添加了分享功能界面,首页热销商品,商品详情页面规格参数界面搭建完成

+++ 天巢网/Classes/home(首页)/controller/RXCollectionController.m
@@ -93,7 +93,7 @@ static NSString * const reuseIdentifier = @"Cell";
 
 #pragma mark <UICollectionViewDelegate>
 - (void)collectionView:(UICollectionView *)collectionView didSelectItemAtIndexPath:(NSIndexPath *)indexPath{
-    NSLog(@"点击了第%ld个item",indexPath.item);
+    NSLog(@"点击了第%ld个item",(long)indexPath.item);
     self.collectionView.userInteractionEnabled = YES;
     TCDetailTableController *vc = [[TCDetailTableController alloc]init];
     vc.view.userInteractionEnabled = YES;
